requirements.txt
requests
pandas
gspread
oauth2client
python-dotenv 